Make keyVInit static
