+++ doc/news/_preparation_next_release.md
@@ -158,7 +158,8 @@ The text below summarizes updates to the [C (and C++)-based libraries](https://w
 - `keyIsDirectBelow` was renamed to `keyIsDirectlyBelow`. _(Philipp Gackstatter)_
 - `keyMeta` was added to provide access to a key's underlying KeySet that holds its metadata keys. _(Philipp Gackstatter)_
 - Removed the obsolete `ksLookupByString` and `ksLookupByBinary`, as well as deprecated `KDB_O_*` options. _(Philipp Gackstatter)_
-- Add `keyLock` and `keyIsLocked`. _(Manuel Mausz)_
+- Added `keyLock` and `keyIsLocked`. _(Manuel Mausz)_
+- Removed `keyVInit`. _(Manuel Mausz)_
 
 ### Opts
